chore: 리뷰 반영
* provider 별 파일로 분리
* axios -> node-fetch 로 변경
* secret key 명칭 변경으로 인한 dotenv 관련 수정
