+++ apps/server/package.json
@@ -26,6 +26,7 @@
         "jsonwebtoken": "^9.0.2",
         "mongoose": "^8.8.1",
         "mongoose-paginate-v2": "^1.8.5",
+        "node-fetch": "^3.3.2",
         "winston": "^3.16.0",
         "winston-daily-rotate-file": "^5.0.0"
     },
