+++ apps/server/src/config/index.ts
@@ -2,6 +2,6 @@ import { config } from 'dotenv'
 
 config()
 
-export const { NODE_ENV, DB_URI, DB_NAME, JWT_SECRET } = process.env
+export const { NODE_ENV, DB_URI, DB_NAME, JWT_PRIVATE_KEY } = process.env
 
 export const PORT = Number.parseInt(process.env.PORT) || 5000
